я ненавижу програмирование

+++ TgCheckerApi/Models/TempScaffold/Channel.cs
@@ -19,7 +19,7 @@ public partial class Channel
 
     public bool? Notifications { get; set; }
 
-    public int? Bumps { get; set; }
+    public decimal? Bumps { get; set; }
 
     public DateTime? LastBump { get; set; }
